Tentative fix for shaped windows (Debian #588537)
Flesh out how we might handle explicitly not applying window shape masks to
windows which would otherwise do so.  Whilst this fixes a bug in display(1)
which is part of Imagemagick, it should be noted that this "fix" will only
apply to windows which -- if they have shaped support defined -- are still
under FVWM's control (that is, without OverrideRedirect set).

In the case of "xteddy -wm" (dear little xteddy!), this will never show a
border around the window, irrespective of any AllowShapeHints style set, as
the client in question never accepts such values in the first place.
There's nothing FVWM can do about this.