Use utils/fvwm-gen-version.sh for version info
When compiling fvwm, we can obtain the version either from Git if it's being
compiled from source, or because it's from a release tarball.

In the case of compiling from Git, then the nearest reachable tag is used,
marking the version as dirty if the index has changes in it.

The output is in a form m4 can understand as this is used directly in
configure.ac.