Add "struct fullscreen" to "struct FvwmWindow"
This introduces an inner-struct to FVWM, which holds different properties of
a window before it goes into fullscreen.  At present, we support restoring
the following attributes of a window:

* Iconic state;
* Layer
* Maximized state;
* Shaded

There may be others in the future.  This struct holds the values of this for
the window, and also stores geometry information for it so that coming out
of fullscreen mode, these values can be referenced.