Correctly indent lists
Since each section when included in a top-level man page is a command
rendered as an ordered list, to ensure paragraphs of that list are indented
correctly, add a literal "+" symbol to each blank line separating out those
paragraphs.

This does not apply to code blocks, where the whitespace is treated
literally.